Turn DashVisualization stateless
